make point module rate limiting more strict
There was some abuse of the bot where a user was repeatedly forwarding
their own messages to cause the bot to impact a whole lot of records at
once and generally annoy the channel. This makes such a thing pointless
as the records aren't updated past the first. Proper, chat-wide rate
limiting is still on the TODO list.