make /sauce show DB info, update DB env var
This commit makes /sauce show what database the bot is currently using.
It also updates the environment variable for setting the database to
match what was written in the README.

Also moves some variable assignments in the source plugin from the
callback to the module, as these only need to be evaluated once and
aren't expected to change as the program is running.